requirements.txt
azure-core==1.5.0
azure-storage-blob==12.3.1
boto3==1.13.6
botocore==1.16.6
cachetools==4.1.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
colorlog==4.1.0
cryptography==2.9.2
docutils==0.15.2
idna==2.9
isodate==0.6.0
jmespath==0.9.5
miniupnpc==2.0.2
msrest==0.6.14
Naked==0.1.31
oauthlib==3.1.0
pathtools==0.1.2
Pillow==7.1.2
protobuf==3.12.2
psutil==5.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycrypto==2.6.1
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
qrcode==6.1
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
s3transfer==0.3.3
shellescape==3.8.1
six==1.14.0
ssdpy==0.2.4
urllib3==1.25.9
